feat: service worker stuff: registered one and then added caching
registered a service worker and created a cache. Then fetch from cache first
and if not available, check network. Verified this is working by killing the
server and seeing if the file still loads. This is done for the main page
only, not yet for the restaurant detail page.

Also, this commit includes some residual experimental stuff in trying to
get aria-setdescendants to work for listbox accessibility.