feat: added service worker/caching to restaurant detail page
when the DOM has loaded, register a service worker for the restaurant detail page (restaurant_info.js)
and share the same service-worker.js file that main.js uses for the restaurant listings page, since
many of the cached items are the same. Each way that restaurant.html is called (with each restaurant id
in the query string) is being cached.